Update UI in a separate thread
If we update UI on for each METAR we process, speed of processing is
limited by ability of terminal to update itself, and because UI updates
very fast it's very flaky.

To solve this, I moved UI updates to a separate thread, which updates
screen on regular intervals.

The only problem I had with this is updating screen when processing is
finished. Because UI updates on intervals, and main thread exits as soon
as it finished processing, UI thread never gets a chance to do one last
update.

To solve this, I've implemented communication between main and UI
threads via MVar. Now UI can properly finish when signaled to stop, and
main is waiting for UI to finish.