Fix average speed reporting
I was using `quot` with `Word` datatype to calculate averages, but
`quot` truncates toward zero, and on large enough dataset it will
inevitably became zero.

This is fixed by using `Double` to keep average, and `/` for division.

I also got rid of `unwindSpeed` which, in addition to making a terrible
pun, also was in a lot of places where it really didn't make sense.